requirements.txt
Werkzeug==0.12.1
Flask==0.12.1
numpy==1.11.3
pandas==0.19.2
pymongo==3.4.0
python_dateutil==2.6.0
